fix: the review of the R3 and R4 core slices, sixteen findings resolved
The settings reader now runs Qt's unescaping state machine over the
whole value, so backslash escapes, hex and octal codes and quoted
sections decode everywhere, lists split only on commas outside quotes,
the @ forms are recognised after unescaping with @@ meaning a leading
@, and keys decode %XX and %UXXXX as code points. Fact metadata refuses
a bitmask index outside the shift range, reads a bool from its text,
and takes mobileDefault on the mobile heads. A component's metadata
keeps its good entries when one is bad. The telemetry log walker
applies the byte-swap rule to every timestamp and resyncs byte by byte
after a bad frame, replay rewinds when played past the end and clamps
a speed that is not positive, the metadata cache keeps an internal
file that is already ahead, the mock link carries its five settings,
and the parameter machine keeps pending writes apart from the facts
so a write cannot mark the load ready, records 64-bit and double
values as unsupported instead of dropping their index, stops only the
initial timer on the PX4 hash check, sends a named refresh for a
component it has not seen, and reports progress in the C++ phases.
Tests no longer sit on incidental data.

+++ core-rs/src/apmmeta.rs
@@ -207,7 +207,7 @@ mod tests {
         assert_eq!(cone.enums.iter().map(|e| e.value.as_f64().unwrap()).collect::<Vec<_>>(), vec![0.0, 1.0, 3.0]);
         let masked = defs.values().flat_map(|m| m.values()).find(|r| r.bitmask.len() >= 3).unwrap().clone();
         let meta = metadata(&defs, 2, &masked.name, ValueType::Int32);
-        assert!(meta.bitmask && meta.enums.iter().all(|e| e.value.as_i64().unwrap().count_ones() == 1));
+        assert!(meta.bitmask && meta.enums.iter().all(|e| (e.value.as_i64().unwrap() as i32 as u32).count_ones() == 1));
         let gain = metadata(&defs, 2, "ATC_RAT_RLL_P", ValueType::Float);
         assert_eq!(gain.decimal_places, Some(6));
         let unknown = metadata(&defs, 2, "NOPE_X1", ValueType::Int32);
